update tms import path
