+++ areal/engine/fsdp_engine.py
@@ -92,7 +92,7 @@ from areal.utils.model import (
     is_qwen_vl_model,
     is_valid_vision_model,
 )
-from areal.utils.offload import is_tms_enabled
+from areal.utils.offload import is_tms_enabled, torch_memory_saver
 from areal.utils.perf_tracer import trace_perf, trace_scope
 from areal.utils.save_load import get_state_dict_from_repo_id_or_path
 from areal.utils.ulysses import (
@@ -101,22 +101,6 @@ from areal.utils.ulysses import (
     ulysses_pad_and_slice_inputs,
     ulysses_prepare_inputs,
 )
-
-try:
-    from torch_memory_saver import torch_memory_saver
-except ImportError:
-
-    class MockTorchMemorySaver:
-        def disable(self):
-            return nullcontext()
-
-        def pause(self):
-            pass
-
-        def resume(self):
-            pass
-
-    torch_memory_saver = MockTorchMemorySaver()
 
 
 @dataclasses.dataclass
